refactor(sso): break up model file into components
