+++ packages/blockchain-wallet-v4-frontend/src/scenes/Login/Wallet/EnterPasswordWallet/index.tsx
@@ -6,28 +6,22 @@ import { bindActionCreators } from 'redux'
 import { Field } from 'redux-form'
 import styled from 'styled-components'
 
-import { HeartbeatLoader, Image, Link, Text } from 'blockchain-info-components'
+import { HeartbeatLoader, Link, Text } from 'blockchain-info-components'
 import { FormError, FormGroup, FormItem, FormLabel, PasswordBox } from 'components/Form'
 import { Wrapper } from 'components/Public'
 import QRCodeWrapper from 'components/QRCodeWrapper'
 import { actions, selectors } from 'data'
 import { LoginSteps, ProductAuthOptions } from 'data/types'
-import { isBrowserSupported } from 'services/browser'
 import { required } from 'services/forms'
 import { isMobile, media } from 'services/styles'
 
 import { Props as OwnProps } from '../..'
-import {
-  ActionButton,
-  BackArrowFormHeader,
-  CenteredColumn,
-  ProductTab,
-  SignUpLink,
-  TabWrapper,
-  UnsupportedBrowserWarning,
-  WalletNeedHelpLink,
-  WrapperWithPadding
-} from '../../model'
+import BackArrowHeader from '../../components/BackArrowHeader'
+import NeedHelpLink from '../../components/NeedHelpLink'
+import ProductTabMenu from '../../components/ProductTabMenu'
+import SignupLink from '../../components/SignupLink'
+import UnsupportedBrowser from '../../components/UnsupportedBrowser'
+import { ActionButton, CenteredColumn, WrapperWithPadding } from '../../model'
 
 const OuterWrapper = styled.div`
   display: flex;
@@ -47,7 +41,6 @@ const SideWrapper = styled.div`
     display: none;
   `};
 `
-
 const FormWrapper = styled(Wrapper)`
   display: flex;
   flex-direction: column;
@@ -57,7 +50,6 @@ const FormWrapper = styled(Wrapper)`
   padding: 0 0 16px 0;
 `}
 `
-
 const MobileAuthSideWrapper = styled(Wrapper)`
   position: relative;
   overflow: visible;
@@ -69,7 +61,6 @@ const MobileAuthSideWrapper = styled(Wrapper)`
   right: 0.5px;
   padding: 40px 16px;
 `
-
 const TextColumn = styled.div`
   display: flex;
   flex-direction: column;
@@ -80,8 +71,6 @@ const TextColumn = styled.div`
   }
 `
 
-const isSupportedBrowser = isBrowserSupported()
-
 const EnterPasswordWallet = (props: Props) => {
   const {
     authActions,
@@ -89,6 +78,7 @@ const EnterPasswordWallet = (props: Props) => {
     formValues,
     handleBackArrowClick,
     invalid,
+    isBrowserSupported,
     qrData,
     submitting,
     walletError
@@ -110,32 +100,18 @@ const EnterPasswordWallet = (props: Props) => {
     <OuterWrapper>
       <SideWrapper />
       <FormWrapper>
-        <TabWrapper>
-          <ProductTab product={ProductAuthOptions.WALLET}>
-            <Image name='wallet-no-background' height='28px' style={{ marginRight: '12px' }} />
-            <Text size='20px' weight={600} color='purple600'>
-              <FormattedMessage id='copy.wallet' defaultMessage='Wallet' />
-            </Text>
-          </ProductTab>
-          <ProductTab
-            backgroundColor='grey000'
-            onClick={onExchangeTabClick}
-            product={ProductAuthOptions.EXCHANGE}
-          >
-            <Image name='exchange-grayscale' height='26px' style={{ marginRight: '12px' }} />
-            <Text size='20px' weight={600} color='grey400'>
-              <FormattedMessage id='copy.exchange' defaultMessage='Exchange' />
-            </Text>
-          </ProductTab>
-        </TabWrapper>
+        <ProductTabMenu
+          active={ProductAuthOptions.WALLET}
+          onExchangeTabClick={onExchangeTabClick}
+        />
         <WrapperWithPadding>
-          <BackArrowFormHeader
+          <BackArrowHeader
             {...props}
             handleBackArrowClick={handleBackArrowClick}
             marginTop='28px'
           />
           <FormGroup>
-            {!isSupportedBrowser && <UnsupportedBrowserWarning />}
+            <UnsupportedBrowser isSupportedBrowser={isBrowserSupported} />
             <FormItem>
               <FormLabel htmlFor='password'>
                 <FormattedMessage
@@ -145,7 +121,7 @@ const EnterPasswordWallet = (props: Props) => {
               </FormLabel>
               <Field
                 name='password'
-                disabled={!isSupportedBrowser}
+                disabled={!isBrowserSupported}
                 validate={[required]}
                 component={PasswordBox}
                 data-e2e='loginPassword'
@@ -193,10 +169,14 @@ const EnterPasswordWallet = (props: Props) => {
                 </Text>
               )}
             </ActionButton>
-            <WalletNeedHelpLink authActions={props.authActions} origin='PASSWORD' />
+            <NeedHelpLink
+              authActions={authActions}
+              origin='PASSWORD'
+              product={ProductAuthOptions.WALLET}
+            />
           </CenteredColumn>
         </WrapperWithPadding>
-        <SignUpLink />
+        <SignupLink />
       </FormWrapper>
       {!isMobile() && (
         <SideWrapper>
