+++ packages/blockchain-wallet-v4-frontend/src/scenes/Login/index.tsx
@@ -14,17 +14,17 @@ import {
   PlatformTypes,
   ProductAuthOptions
 } from 'data/types'
+import { isBrowserSupported } from 'services/browser'
 
-// step templates
 import Loading from '../loading.public'
 import MergeAccountConfirm from './AccountUnification/MergeAccountConfirm'
 import UpgradePassword from './AccountUnification/UpgradePassword'
 import UpgradeSuccess from './AccountUnification/UpgradeSuccess'
+import LoginSceneFooter from './components/LoginSceneFooter'
 import ExchangeEnterEmail from './Exchange/EnterEmail'
 import EnterPasswordExchange from './Exchange/EnterPasswordExchange'
 import InstitutionalPortal from './Exchange/Institutional'
 import TwoFAExchange from './Exchange/TwoFA'
-import { getLoginPageFooter } from './model'
 import { getData } from './selectors'
 import VerifyMagicLink from './VerifyMagicLink'
 import CheckEmail from './Wallet/CheckEmail'
@@ -118,6 +118,7 @@ class Login extends PureComponent<InjectedFormProps<{}, Props> & Props, StatePro
       busy,
       exchangeError,
       handleBackArrowClick: this.handleBackArrowClick,
+      isBrowserSupported: isBrowserSupported(),
       isMobileViewLogin: platform === PlatformTypes.ANDROID || platform === PlatformTypes.IOS,
       setStep: this.setStep,
       walletError,
@@ -127,7 +128,6 @@ class Login extends PureComponent<InjectedFormProps<{}, Props> & Props, StatePro
     return (
       <>
         {/* CONTENT */}
-
         <Form onSubmit={this.handleSubmit}>
           {(() => {
             switch (step) {
@@ -163,7 +163,7 @@ class Login extends PureComponent<InjectedFormProps<{}, Props> & Props, StatePro
         </Form>
 
         {/* FOOTER */}
-        {!loginProps.isMobileViewLogin && getLoginPageFooter(step)}
+        {!loginProps.isMobileViewLogin && <LoginSceneFooter step={step} />}
       </>
     )
   }
@@ -193,11 +193,12 @@ const mapDispatchToProps = (dispatch) => ({
 
 const connector = connect(mapStateToProps, mapDispatchToProps)
 
-type FormProps = {
+type OwnProps = {
   busy?: boolean
   exchangeError?: ExchangeErrorCodes
   handleBackArrowClick: () => void
   invalid: boolean
+  isBrowserSupported: boolean
   isMobileViewLogin?: boolean
   pristine: boolean
   setStep: (step: LoginSteps) => void
@@ -208,7 +209,7 @@ type FormProps = {
 type StateProps = {
   captchaToken?: string
 }
-export type Props = ConnectedProps<typeof connector> & FormProps
+export type Props = ConnectedProps<typeof connector> & OwnProps
 
 const enhance = compose<any>(reduxForm({ form: LOGIN_FORM }), connector)
